+++ RedPandaIDE/utils.cpp
@@ -586,7 +586,7 @@ void stringsToFile(const QStringList &list, const QString &fileName)
     }
 }
 
-void resetCppParser(std::shared_ptr<CppParser> parser)
+void resetCppParser(std::shared_ptr<CppParser> parser, int compilerSetIndex)
 {
     if (!parser)
         return;
@@ -600,7 +600,10 @@ void resetCppParser(std::shared_ptr<CppParser> parser)
     parser->setParseLocalHeaders(true);
     // Set options depending on the current compiler set
     // TODO: do this every time OnCompilerSetChanged
-    Settings::PCompilerSet compilerSet = pSettings->compilerSets().defaultSet();
+    if (compilerSetIndex<-1) {
+        compilerSetIndex=pSettings->compilerSets().defaultIndex();
+    }
+    Settings::PCompilerSet compilerSet = pSettings->compilerSets().getSet(compilerSetIndex);
     parser->clearIncludePaths();
     if (compilerSet) {
         foreach  (const QString& file,compilerSet->CppIncludeDirs()) {